Parse Board State in Python
This allows for basic querying about the state of the board from within
the Python program. This includes asking who's turn is it and what piece
is located at a given board position.

Currently this is only implemented for Checkers, but now that the
structure exists, it shouldn't be too difficult to implement it for
Othello and C4Pop10 as well.