Stash valid moves
Running regex to parse arbitrary move strings is expensive. Since in most
cases the program will just be feeding the same strings generated by
getValidMoves back into doMove, it saves time to  cache the generated moves
and use relatively quicker string comparison to lookup which move to
apply. Then, we only have to regex-parse the move string in the case that
the move is not already in the cache.