Fix unsigned char issue
By spec, it is up to the compiler to decide whether or not the char data
type is signed. Since most compilers assume it is signed, I was using the
char data type to hold signed eight-bit values. However, this posed an
issue when I compiled on my raspberry pi and found out that it defaults
char to be unsigned. To fix this, I had to use "signed char" instead of
just "char" everywhere where the sign was relevent.